census: the comment stripper was deleting code, and it invented four findings
⚠️ THE INSTRUMENT, AGAIN, AND IN THE OTHER DIRECTION.

Stripping comments before looking for reads was right — this file's own header
had been counting as a read of the fields it names as examples. The
implementation was not. `/\*[\s\S]*?\*\//g` treats the `/*` inside a string
literal as the start of a block comment and runs to the next close, so
`app.all("/api/platform/*", …)` in `workers/app.ts` opened one that swallowed
6108 of that file's 15732 characters. Thirty-nine per cent of a file, silently
deleted from the corpus, taking a real `.hasPortalIntegrationApi` read with it
and reporting that field as unread.

Deleting code is the same error as counting prose, in the same direction: it
invents findings. Four of them, all now dropped —
`DeploymentProfile.hasPortalIntegrationApi`, `CancellationPolicy.remedy`, and
both `colour` fields, two of which were next on the list to be investigated as
real gaps.

A comment is now recognised only when it OPENS ITS OWN LINE, block or line
alike. That is exact for this codebase's style and cannot be fooled by a path
inside a string.

TWELVE ENTRIES CLASSIFIED, on evidence:

  wire-shape (7) — values that arrive from or are stored for somebody else.
    `GoogleTokenResponse.token_type` (the provider sends it; Intuit's schema has
    the same field), `ComplianceEvent.entitySid` (Telnyx's and Twilio's own
    identifier, ours to pass on), `MigrationBundleV1.formatVersion` (a stored
    discriminator all three adapters write), both `photosEmbedded` manifests
    (they cross the seam and portal stores them against the DSAR record),
    `ProductionInput.producedBy` (written to `statutory_productions.produced_by`)
    and `OutboxRow.lastTriedAt` (mirrors its column).

  invisible-read (1) — `AuthVariables.requestedTenantSlug` is read as
    `c.get('requestedTenantSlug')` in three routes. The key is a string.

  deferred (4), each a small thing that does nothing:
    `IntegrationRow.lastSync` — every integration row carries when it last
      synced and no surface shows it. The calendar badge reads a DIFFERENT field
      (`calendarLastSyncAt`), so the page that lists every connection is the one
      place holding the number and not using it.
    `AiRequest.topK` — a sampling parameter the request shape accepts and the
      single adapter never sends, so a caller can set it and nothing happens.
    `AuditActionDef.altFamilies` — an action's alternative audit families,
      declared with a comment explaining the check they are for. That check does
      not run.
    `AgentTermsVersions.requiredContentHash` — set on every branch and read by
      nobody, so an agent told they must re-accept is not told which version.

Still owed: 51 -> 39.

The declaration parser moves to `scripts/lib/type-properties.mjs`: this gate
crossed the 400-line cap as a NEW file over it rather than a grandfathered one,
so the rule asks for an extraction. "List the declared properties of these
sources" is worth answering on its own anyway.
