+++ scripts/check-unread-fields.mjs
@@ -73,6 +73,7 @@ import { execFileSync } from 'node:child_process';
 import { readFileSync, writeFileSync, existsSync } from 'node:fs';
 import { dirname, join, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
+import { declaredProperties } from './lib/type-properties.mjs';
 
 const ROOT = resolve(dirname(fileURLToPath(import.meta.url)), '..');
 const BASELINE = join(ROOT, 'scripts', 'unread-fields-baseline.json');
@@ -157,77 +158,43 @@ const READ_ONLY_CORPUS = execFileSync('git', ['ls-files', '--cached', '--others'
  * disappeared for no reason but that. A comment mentioning `.foo` HIDES a real
  * finding, which is the expensive direction to be wrong in.
  *
- * Block comments go entirely — that is where prose lives. Line comments go only
- * when the whole line is one, so a `https://…` inside a string is never mistaken
- * for the start of a comment.
+ * A comment is recognised only when it OPENS ITS OWN LINE, block or line alike.
+ * The obvious regex form is worse than no stripping at all: a string literal
+ * like `"/api/platform/*"` opens a block comment that runs to the next close,
+ * and it ate 6108 of `workers/app.ts`'s 15732 characters — 39% of a file,
+ * silently, taking a real `.hasPortalIntegrationApi` read with it and reporting
+ * that field as unread. Deleting code from the corpus is the same error as
+ * counting prose, in the same direction: it invents findings.
  */
 function withoutComments(src) {
-    return src
-        .replace(/\/\*[\s\S]*?\*\//g, ' ')
-        .split('\n')
-        .map((line) => (line.trimStart().startsWith('//') ? '' : line))
-        .join('\n');
+    const out = [];
+    let inBlock = false;
+    for (const line of src.split('\n')) {
+        const t = line.trimStart();
+        if (inBlock) {
+            if (line.includes('*/')) inBlock = false;
+            out.push('');
+            continue;
+        }
+        if (t.startsWith('/*')) {
+            if (!line.includes('*/')) inBlock = true;
+            out.push('');
+            continue;
+        }
+        out.push(t.startsWith('//') ? '' : line);
+    }
+    return out.join('\n');
 }
 
 const READERS = new Map([
     ...[...SRC].map(([f, src]) => [f, withoutComments(src)]),
     ...READ_ONLY_CORPUS.map((f) => [f, withoutComments(readFileSync(join(ROOT, f), 'utf8'))]),
 ]);
-
 /* ------------------------------------------------------------------ */
 /*  Declarations                                                       */
 /* ------------------------------------------------------------------ */
 
-/**
- * The body of an object type, found by matching braces from the declaration's
- * opening `{`.
- *
- * The obvious line-based approximation — start at the declaration, add braces,
- * subtract braces — walks straight off the end of the type and into the file,
- * and then reports every function PARAMETER below it as a property of that
- * type. Measured while writing this: it invented ten fields on two types in
- * `app/lib/editor/structure-ops.ts`, all of them parameters of the functions
- * that happened to follow.
- */
-function bodySpan(src, openBraceIdx) {
-    let depth = 0;
-    for (let i = openBraceIdx; i < src.length; i++) {
-        const ch = src[i];
-        if (ch === '{') depth++;
-        else if (ch === '}') {
-            depth--;
-            if (depth === 0) return [openBraceIdx + 1, i];
-        }
-    }
-    return null;
-}
-
-const DECL = /(?:export\s+)?(?:interface\s+(\w+)\s*(?:extends\s+[^{]+)?|type\s+(\w+)\s*=\s*)\{/g;
-/** `name?: …` / `readonly name: …` at the top level of a type body. */
-const PROP = /^[ \t]*(?:readonly\s+)?(\w+)\??\s*:/;
-
-const declared = [];
-for (const [file, src] of SRC) {
-    for (const m of src.matchAll(DECL)) {
-        const open = m.index + m[0].length - 1;
-        const span = bodySpan(src, open);
-        if (!span) continue;
-        const typeName = m[1] ?? m[2];
-        const body = src.slice(span[0], span[1]);
-        const firstLine = src.slice(0, span[0]).split('\n').length;
-
-        let depth = 0;
-        body.split('\n').forEach((line, i) => {
-            const atTop = depth === 0;
-            depth += (line.match(/\{/g) ?? []).length - (line.match(/\}/g) ?? []).length;
-            if (!atTop) return;
-            const trimmed = line.trim();
-            if (trimmed.startsWith('*') || trimmed.startsWith('//') || trimmed.startsWith('[')) return;
-            const p = line.match(PROP);
-            if (p) declared.push({ file, type: typeName, prop: p[1], line: firstLine + i });
-        });
-    }
-}
+const declared = declaredProperties(SRC);
 
 /* ------------------------------------------------------------------ */
 /*  Reads                                                              */
